fix: remove unnecessary focus clearing on shape add/delete
Whenever shapes were added or deleted, the focus of the selection would
be cleared once as a precaution against outdated shapeIds remaining in
the selectedShapeIds. This interfered with programmatic addition and
selection of fields, though. In practice, the default behavior of
the focus when deleting and adding fields is enough to reset the
selectedShapeIds in a healthy way, anyway.