feat: use minimumDistanceForSelection to avoid extra selection on click
When a UI built on the shape editor assigns extra behavior to click
methods, the unconditional calling of selection rectangle behavior can
cause unintentional side effects. Therefore, minimumDistanceForSelection was
added with a reasonable default of 15 pixels to avoid starting a
selection until that many pixels have been spanned relative to the
starting point of the selection mousedown.